also reset transport before tests
Previously the transport was only reset after tests in a testcase using this
trait were run. This can cause an issue though if other tests
run first which send messages, but do not use this trait. In
this case the transport isn't cleared so the first test to run in a test
case *with* this trait will have the state from the previous tests.